refactoring: moved auth and email to domain
