+++ internal/domain/auth/error.go
