+++ internal/domain/email/dto.go
