+++ internal/domain/email/entity.go
