+++ internal/domain/email/error.go
