+++ internal/domain/email/service.go
