+++ internal/handler/http/v1/handler.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	"github.com/answersuck/vault/internal/app/auth"
 	"github.com/answersuck/vault/internal/config"
+	"github.com/answersuck/vault/internal/domain/auth"
 
 	"github.com/answersuck/vault/pkg/logging"
 )
